[libromdata] GameCube: Detect more types of standalone update partitions.
'UPD' is an IOS-only update partition.
'UPE', 'UPJ', 'UPK', and 'UPP' are region-specific update partitions.

'UPC' would be a region-specific update partition for China, if the
iQue Wii would have been released.

There might be a 'UPT' or 'UPW' for Taiwan, but a Taiwan-region Wii
didn't happen, either. (Well, there might be an RVL-001(TWN), but it
likely has a JPN region code.)