[win32] ConfigDialog: Restore the default system menu.
Property sheets only have "Move" and "Close" by default. This re-adds
all of the other entries.

In addition, this seems to fix something that broke on Windows 8.1
and later (and possibly Windows 8.0): Re-adding the Minimize button
without restoring the system menu results in a Minimize button that
does absolutely nothing when clicking it.

This fixes issue #229: Minimize button doesn't work on win32 rp-config