[rp-download] IDownloader: Added an os-release parser to get the Linux distribution.
If the NAME= field doesn't have "Linux" it's added afterwards.
If not found, it defaults to "Linux".

Added a semicolon between the OS name and CPU name for all systems.
This was already done for Windows, but I left it out of the other ones to
match Firefox. (Firefox should probably add the semicolon...)