[libromdata] NintendoDS: Indicate the specific Security Data that's present.
There's two important sections: Blowfish Tables and Static Data.
The rest is some randomized data. The exact algorithm is unknown,
and ROM images work on an IS-NITRO-EMULATOR if Blowfish and Static
data are present but Random data isn't. Missing Random data does
cause a verification error when using the official devkit tools
to verify a flashcart, so that means there's *some* algorithm used
to calculate it.