[libromdata] NintendoDS: Implemented Trim/Untrim.
UI frontends: Run the ROM operations. The menu label will be updated
afterwards.

TODO: Need to improve UI feedback to indicate that the operation
completed. We could use KMessageWidget on KDE, but I'm ont sure
what to do on GTK+ and Win32. Maybe at least a message beep on
Windows, and other audio feedback on the other systems...