[kde] RomDataViewPrivate::updateField(): Added support for RFT_BITFIELD.
bitfield_toggled_slot() -> bitfield_clicked_slot()

The "toggled" signal is emitted whenever the checkbox state is changed
by either the user *or* by code, whereas "clicked" is only emitted when
it's changed by the user. We're changing the checkbox state in
updateField(), so we need to use "clicked", not "toggled"; otherwise,
it won't actually take effect.

(...actually, it *will* take effect, if the user clicks the box, since
we're setting the new value as "RFT_BITFIELD_value".)

Encrypt/Decrypt now updates both fields correctly on KDE.