[kde] RomDataView: Use an std::unordered_map<> to map field indexes to QObjects.
The previous method would have needed a lot of special cases in order
to handle everything properly.

This method does use some extra memory, but for e.g. a RomData object
with 100 fields, it would take up around 100 * (4 + 8) == 1,200 bytes
for data, plus std::unordered_map<> overhead, so it's not that much.