[win32] RP_ShellPropSheetExt: Initial updateField() implementation.
Based on the KDE version.

RFT_STRING is implemented; RFT_BITFIELD is not (yet).

We don't need an std::unordered_map<> to map IDs to controls because the
controls all have dialog IDs based on the field indexes, and we can use
GetDlgItem() to retrieve a control based on a dialog ID.

Other changes:
- createHeaderRow(): s/tSysInfo/ts_sysInfo/