[kde] RomDataView: Added a KMessageWidget for the ROM operations status message.
The KF5 version has slots indicating both the Show and Hide animations
finished, so we use that to control the QTimer.

The KDE4 version does not have these, so we start the timer when the
KMessageWidget is set up. We can't easily stop it if the Close button
is clicked, but that won't do anything, since it'll try hiding the
KMessageWidget when it's already hidden.