[gtk] MessageWidget: Added background and border colors for GTK+ 2.x.
The base class is now GtkEventBox on GTK+ 2.x, since GtkEventBox allows
setting the background color, but GtkHBox does not.

An inner GtkEventBox allows for setting a border color, and then finally
a GtkHBox provides the original container.

The lighten() function isn't available here, so I'm hard-coding the same
colors that were used on Thunar on GTK+ 3.x here.