[rp-download] rp-download.cpp: Add an extra fflush() just in case something screws up.
When trying to reimplement open-by-handle for Windows ADS, I noticed
the file was being truncated on a 4 KB boundary. It seems this extra
data isn't written until the file is closed and/or flushed, and there
was something screwy with NtCreateFile() that was messing it up.

This doesn't happen with the current code, but add an fflush() anyway.