[libromdata] WiiWAD: Added preliminary support for Nintendo DSi TAD packages.
These have the same basic format as WADs, including ticket and TMD, but
the contents are different.

Basically, all the stuff that's not useful in DSi TMDs has been hidden.
Region code and age ratings are part of the SRL, which I'll add in a
separate tab at a later point.