[libromdata] NintendoDS: Rewrote animated icon handling to support the flip bits.
KADJ_SPD-02799.tad uses H-flip and V-flip for its counting digits, so
we need to implement flip in order for '5' and '6' to show up correctly.

There's still a maximum of 64 frames, but we're using a mapping from
the token's high byte to the frame index. That way, we don't have to
reserve 256 frames instead.

[librptexture] rp_image_ops.cpp: vflip() didn't copy the palettte for
CI8 images. It does now.

TODO: Implement H-flip.