[libromdata] WiiWAD: Added SRL extraction.
Split the private class declaration into WiiWAD_p.hpp.

FIXME: Extracting an SRL with an animated icon causes the program to
crash. This is because the rp_images objects are deleted when the SRL
is reopened, which results in the UI holding onto dangling pointers.

Looks like it's time to make rp_image a refcounted object...