[librptexture] rp_image: Changed to a reference-counted class.
This should fix issues when the main content object is deleted and
reopened in order to extract e.g. an SRL.

Added special const rp_image::ref() and const rp_image::unref()
overloads to allow ref() and unref() of const rp_image objects,
since this is used everywhere.

[libromdata] ImageDecoderTest: Use a custom deleter for unique_ptr<rp_image*>.

In general, I'd rather not use unique_ptr<> with refcounted objects,
since that causes confusion with the reference count, but for this
particular case, it should be fine.