UI frontends: Take references to the rp_image objects.
This seems to fix the crash on Linux, though it crashes with an
assertion on Windows.

On Linux, the animation stops. I probably need to make IconAnimData a
refcounted object.

NOTE: On Windows, it crashes because the MSVC debug runtime intentionally
sets freed memory to a specific value instead of leaving it undefined.
On Linux, set can set MALLOC_CHECK_=3 and MALLOC_PERTURB_=3 for similar
results, in which case, it crashes with an assertion.