[win32] Set WS_EX_LAYOUTRTL if the system is using an RTL language.
This is needed for proper right-to-left layout on systems that use an
RTL language, e.g. Hebrew.

We're determining if the system is RTL by checking if the taskbar, aka
Shell_TrayWnd, has WS_EX_LAYOUTRTL. Windows Explorer on Windows 7's
GetProcessDefaultLayout() returns 0, even if an RTL language is in use.

rp-config: WS_EX_LAYOUTRTL is needed for proper positioning of the dialog
buttons. Otherwise, the custom buttons either overlap or get overlapped
by the standard buttons. Also, this reverses the TreeView layout.