[win32] RP_ShellPropSheetExt: Flip ImageList images horizontally if using an RTL layout.
ImageList and/or the individual controls flip the images by default.
There's an ILC_MIRROR option, but it's not available in Win2000 and
earlier, and it depends on GetProcessDefaultLayout(), which isn't
always set correctly.

This fixes flags and Xbox 360 achievement icons, among other things,
when using an RTL language on Windows.