Several fixes found using PVS-Studio 7.09.41421.64.
[kde] RomDataView.cpp: V547 (Med): Expression 'name' is always true.

[kde] SystemsTab::loadDefaults(): Fixed a typo that caused the DMG
dropdown to be set to the CGB value, and the CGB dropdown to not be
reset at all.

V525 (Low): The code contains the collection of similar blocks.
Check items 'cboDMG', 'cboSGB', 'cboDMG' in lines 134, 138, 142.

[libromdata] Use hexadecimal values instead of multi-character literals
if an escape character is needed for one of the bytes.

V1039 (High): Character escape is used in multicharacter literal. This
causes implementation-defined behavior. Inspect expression: ''GBS\x01''.

[libromdata] Xbox360_XEX_Private::getXdbfResInfo(): Initialize res.size.

V614 (Med): Potentially uninitialized variable 'res.size' used.

[libromdata] Nintendo3DSPrivate::openSRL(): Use UNREF() in case
srlData is nullptr.

V522 (Low): There might be dereferencing of a potential null pointer 'srlData'.

[libromdata] PSPPrivate::openBootExe(): Check exeData->isOpen().

V668 (Med): There is no sense in testing the 'exeData' pointer against
null, as the memory was allocated using the 'new' operator. The exception
will be generated in the case of memory allocation error.