[win32] RP_ShellPropSheetExt: Handle WM_CTLCOLORSTATIC in the subtab dialog procedure.
Handling it in the main dialog procedure only worked before if there were
no subtabs. Since we changed it to use a child dialog even in cases where
we don't have subtabs (for WS_VSCROLL), we need to handle it in the subtab
dialog procedure instead.

Technically, we should have been handling it in the subtab dialog
procedure in the first place, but STRF_WARNING was never used in cases
where more than one subtab were present.