[win32] RP_ShellPropSheetExt: Handle mousewheel scrolling on tabs with WS_VSCROLL.
Note that the parent window might receive mousewheel messages in some
cases, so those messages are forwarded to the child window.

TODO: Remember which control had focus on each tab and set focus to
that control. For now, we're not adjusting focus, so switching tabs
might cause wheel movement to affect the *previous* tab.