[rpcli] Added ATA IDENTIFY PACKET DEVICE.
Not entirely sure if it works properly, but I had this in my `git stash`
listing for a while, so I might as well commit it.

On VirtualBox, it returns garbage data for the emulated CD-ROM drive,
but so does ATA IDENTIFY DEVICE...

FIXME: Also on VirtualBox, it failed on Windows if no disc was present.