[libromdata] MachO: Added 64_32 checks. (based on `file`)
I'm not entirely sure what the 64_32 ABI is, but I'm assuming it's
similar to the little-used Linux x32 ABI.

MachOData: The 64-bit CPU type table is mostly nullptr, so get rid of
it and use a small switch/case. This results in a smaller object file,
even with the 64_32 additions.

Code size reduction: (64-bit Gentoo Linux, gcc-10.2.0, release build, no LTO)

Without 64_32 additions:
   text    data     bss     dec     hex filename
   1856     624       0    2480     9b0 MachOData.cpp.o [before]
   1904     472       0    2376     948 MachOData.cpp.o [after]
    +48    -152       0    -104     -68 Difference

With 64_32 additions:
   text    data     bss     dec     hex filename
   1856     624       0    2480     9b0 MachOData.cpp.o [before]
   1945     472       0    2417     971 MachOData.cpp.o [after]
    +89    -152       0     -63     -3f Difference