[librpbase] RpPng: Set a fake sBIT if the original PNG doesn't have one.
This is definitely needed for IThumbnailProvider, since the sBIT is used
to determine if the image is transparent. That was a last-minute change
in v1.7, and I don't think it's that important that an update needs to
be pushed.

This only affects Xbox 360 and PlayStation Portable, since those formats
use PNG images internally, and only if the PNG images don't have an sBIT
chunk. (I'm not sure how many of them do.)