[win32] AchWin32: Use a separate window per thread, and destroy the icon in WM_NCDESTROY.
On Windows 7, the file properties dialog has its own thread. Closing the
dialog terminates the thread, which causes our hidden window to be closed.
We'll need to destroy the tray icon when this happens; otherwise, it'll
be left dangling until the user moves their mouse over it.

The hidden window now has a window procedure. The notification icon will
send notifications using WM_ACHWIN32_NOTIFY.

TODO: Handle notifications, e.g. if the user closes the popup, and/or
automatically close the hidden window after a timeout.