[kde] AchQtDBus: Use the custom achievement icons.
Based on the same code from the Windows version, though we're using
QImage instead of rp_image.

QDBusArgument operators are used to (de)serialize the NotifyIconStruct
for D-Bus, since we have to put it in a QVariantMap.

NOTE: The hint name for the image changed in various versions of the
specification. We're using the oldest version for compatibility purposes.

CMakeLists.txt: Build ach.qrc. We added the file before, but it wasn't
actually compiled in.