AchGDBus, AchQtDBus: Escape the notification summary.
KDE didn't seem to have a problem with an unescaped '&', but XFCE
(GTK+ 2.x) refused to include the unescaped summary in the notification.
I'm not sure about GTK+ 3.x, but I suspect it's the same as GTK+ 2.x.

Escape it on both KDE and GTK+ systems. There's effectively no change on
KDE, but this fixes the summary on GTK+.