[libromdata] GameCube: Check for incrementing values in debug-encrypted partitions.
Moved the check to WiiPartition. When we initialize decryption for
encrypted discs, if the Wii Magic isn't present, check for incrementing
values.

This fixes detection of incrementing values on encrypted versions of
Wii Menu Installer v1.

Fixes #269: Keys are correct but property sheet still reports that the
key is incorrect.
- Reported by @Masamune3210.