[librpbase] RomFields: RFT_LISTDATA: Add column sizing parameters.
This allows us to specify which columns should be expanded, and which
columns should be autosized.

Note that this is only fully supported on KDE. GTK+ supports some of
the functionality, and for the current use case, it's basically
equivalent to KDE.

Win32 ListView doesn't support this at all, unless I do some custom
control coding at some point.

[libromdata] Xbox360_XDBF: Add colsz values. ID and Gamerscore fields
should be autosize; Description should be 'stretch'.