[kde] LanguageComboBox::setLCs(): Don't call count() repeatedly.
We're clearing the QComboBox, so we can keep track of the current index
here instead of calling count().

Use selectedLC() to get the selected LC. This is similar to the new
GTK+ LanguageComboBox code I'm about to commit. Note that this might
reduce performance *slightly*, since it's no longer inlined.