[win32] New multi-language ComboBoxEx superclass.
Note that this is considered a "superclass" by Microsoft, since a
subclass doesn't have its own window class.

https://docs.microsoft.com/en-us/windows/win32/winmsg/about-window-procedures

This uses a 0-terminated lcs_array, similar to the GTK+ version.

FIXME: WM_LCB_GET_LCS isn't implemented right at the moment. Returning
a malloc()'d pointer in response to a window message is not a good idea.