[gtk] RomDataView: Combine the icon/checkbox and column 0.
We can pack multiple GtkCellRenderers into a single column, so do that
in order to get rid of the separate icon/checkbox column.

Note that this does mean the GtkListStore column IDs will be out of sync
with the GtkTreeView column IDs, so we have to take that into account
when updating things.