[amiibo-data] New amiibo database binary format.
amiibo-bin.txt: Description of the binary format.
amiibo-data.txt: Amiibo database in text format.

The database will be compiled to a binary file using amiiboc.

TODO: CMake's cross-compilation functionality is a bit iffy and requires
building a host system version of amiiboc in a separate CMake run. Maybe
we should convert it to Perl or Python?