[gtk] OptionsMenuButton: Preliminary support for GMenuModel.
GMenuModel was added in glib-2.32 and GTK+ 3.4.

GTK+ 4.x removes GtkMenu and GtkMenuItem entirely, so we need to start
preparing for that change.

FIXME: We can't handle menu item state or signals unless we set up
GActions, which are apparently application-wide. We'll need to create
temporary action names to avoid problems.

The GMenuModel code is currently disabled until everything is working.