[win32] rp-es_ES.rc: Initial Spanish Win32 resource script.
Moved the ImageTypesTab credits text out of the UI and RC files.
The Spanish version was >256 characters, which is truncated by the
MS resource compiler. (RC4206)

The string is stored with only "\n". The KDE frontend converts it to
"<br/>" for rich text display purposes.

AboutTab: Removed the es translation of the program name, since it
wasn't properly translated from Portuguese.