[kde] Initial implementation of CacheTab.
The caches aren't cleaned yet, but the infrastructure is there.

CacheCleaner runs in a separate thread and sends status messages using
Qt signals.

RpProgressBar: QProgressBar subclass that supports setting an "error"
display.

NOTE: If a sound is played and the program is exited before the sound
is finished, something crashes in Phonon::Gstreamer. This isn't an
issue with rom-properties, so I can't really do much about it.

[libunixcommon] userdirs.cpp: Make isWritableDirectory() public.