[librpbase] uvector.h: Fix build on gcc-11.
gcc-11 defaults to C++ 2017 instead of C++ 2014, so a new code path
is used. In this case, `allocator_type` is defined later on, so we have
to rearrange the code such that `allocator_type` is defined before we
attempt to use it.

Also, we have to add `typename` because of weird stuff with
`allocator_type`.