[kde] RomDataView: Use QLocale::system().toString() for RFT_DATETIME.
This is available starting with Qt 4.4.

Qt::DefaultLocaleShortDate and related are deprecated and will be removed
in Qt 6:

../../src/kde/RomDataView.cpp: In member function ‘void RomDataViewPrivate::initDateTime(QLabel*, const LibRpBase::RomFields::Field&, int)’:
../../src/kde/RomDataView.cpp:885:60: warning: ‘Qt::DefaultLocaleShortDate’ is deprecated: Use QLocale [-Wdeprecated-declarations]
  885 |                         str = dateTime.date().toString(Qt::DefaultLocaleShortDate);
      |                                                            ^~~~~~~~~~~~~~~~~~~~~~

NOTE: I was going to use QDateTime::toString(locale.dateTimeFormat()),
which is available starting with Qt 4.1, but this won't be localized
as of Qt 6. Our minimum Qt 4 version is 4.8, so the QLocale-only version
will work better.