[kde] OptionsTab: Added an option for setting the language to use for PAL titles on GameTDB.
[librpbase] Config: Likewise.

The option is saved as the two-character language code.

RpQt.hpp: New function lcToQs() that converts a language code
to a QString. This is used in a few places.

LanguageComboBox: New setLCs() overload that takes an array of
language codes and repacks them into an std::set<> for processing.

TODO:
- Windows version.
- Actually make use of this setting when downloading.