[libromdata] GameCube, WiiU, WiiWAD, Nintendo3DS, NintendoDS: Use the PAL fallback language code.
The functions now return vector<uint16_t>, which is a GameTDB language
code. Note that it might need to be converted to uppercase, since the
language code stored in the config file is lowercase, whereas GameTDB
uses uppercase language codes.

Note that even though it's called "fallback", it's actually used
*first* for e.g. games with region character 'P'. Maybe it should be
renamed to "preferred" PAL language...