[libromdata] MegaDrive: Improve "Game Toshokan" handling.
No-Intro's "SegaNet" ROMs have a Game Toshokan boot ROM (128 KB) and a
downloaded game (128 KB). The "SegaNet" versions usually have a fixed
checksum, but the "Game no Kanzume Otokuyou" versions do not.

Check for the Game Toshokan serial number. If it's found, and the ROM
is 256 KB, calculate the CRC32 of $20000-$200FF and use that as the
ROM serial number in the "GT" region.

This allows us to thumbnail all of the known Game Toshokan ROM images
in No-Intro.