[libromdata] MegaDrive: Handle 'early' ROM headers.
I'm not sure if it's actually an early format or just one batch of
broken ROMs, but the following ROMs have 32-byte titles:

- Fatman (Japan)
- Sega Sound Tool (USA) (v2.2) (Program)
- Slaughter Sport (USA)

Fatman and Slaughter Sport are the same game but with localized titles.
We can now add title screens for these ROMs because the serial number is
identified properly.

Sega Sound Tool has a generic serial number (GM XXXXXXXX-XX), so we can't
add a title screen yet.