[amiibo-data] amiibo-data.txt: Added Super Nintendo World Power-Up Bands.
The amiibo database is now up to date as of 2021/06/07.

Added NFP_TYPE_BAND for the power-up band type.

NOTE: The amiibo series is 0xFF. Since we're using a plain array for
amiibo series names in the database file, this results in lots of zero
entries. It's not worth changing this to save less than 1 KB, though.