[tinyxml2] CMakeLists.txt: Fix DLL installation in Windows packaging, again.
This was originally fixed in commit b243f47341b7997c6fae7167d73736032f17c7ce.
([tinyxml2] Fix DLL installation in Windows packaging.)

However, it regressed when TinyXML2 was upgraded to 8.1.0 in
commit 05aed01ddf598e21135cb64284cb26cad6abad97.
([tinyxml2] Updated to tinyxml2-8.1.0.)

The end result was the 32-bit TinyXML2 being installed in bin/,
and the 64-bit TinyXML2 not being installed at all. This resulted in
crashes, which shouldn't have happened either; this means something
is wrong with DelayLoad, so that will need to be fixed next.

Fixes #313: tinyxml2.dll isn't packaged correctly
Reported by @ccawley2011.