[gtk] Initial experimental support for GTK4.
I don't have any GTK4 file browsers to test this, but it compiles.
GTK4 is disabled by default due to this.

GTK4 doesn't support Cairo, so we're falling back to GdkPixbuf for
now. There *is* a new GdkTexture class, which implements most of
GdkPixbuf's functionality without the overhead as Cairo.

Missing functionality:
- Drag & drop needs to be completely rewritten.
- gtk_dialog_run() doesn't exist anymore, so file dialogs need to be
  reworked to use a signal handler. This will work in older versions
  of GTK, too.
  - Also: Handle URIs properly?
- GtkHeaderBar: GTK4 no longer has GtkButtonBox. Figure out "secondary"
  there.
- GTK4 equivalent of gtk_container_set_border_width().

The Nautilus frontend is enabled, but Thunar is not, since that requires
more changes that can't easily be verified at the moment.