[gtk] RomDataView: Add support for Nautilus 40, which uses libhandy.
Instead of linking to libhandy-1, use dlvsym() to dynamically load the
three required functions. These functions will already be loaded in the
process if Nautilus uses libhandy-1, so we don't have to dlopen() the
shared library. (Use RTLD_DEFAULT as the module handle.)

dlvsym() is used for symbol versioning if available. (LIBHANDY_1_0)

Added a simple wrapper function gtk_widget_get_first_child() for
GTK+ 3.x and earlier, which assumes the specified widget is a
GtkContainer.