[libwin32common] WinUI: Revert OFN_NOCHANGEDIR for now.
This was accidentally committed with the CPU macro cleanup. It was based
on an Old New Thing article I read about this macro:

https://devblogs.microsoft.com/oldnewthing/20101112-00/?p=12293

I wanted to test it to verify that it had the right behavior, but I forgot
to do that before committing. Either way, it somehow broke the AppVeyor
build because the macro is being changed to "OFN_NO-CHANGE-DIR", and I
have no idea why.