[win32] RP_ExtractIcon_Fallback: Remove the Application fallback.
Since we're no longer registering under Applications, this is no longer
needed.

If a UserChoice application is selected, the icon will now be the
program icon instead of the rom-properties thumbnail. The provided
thumbnail via RP_ThumbnailProvider will still be correct, and the
program icon will appear in the bottom-right.

This is probably the best option we have here, since the other option
is "breaking randomly".