[cmake] win32-gcc.cmake: Use the _ld_out method instead of assuming ld always supports these options; add more options.
Previously, it tested for --large-address-aware, --nxcompat, and
--tsaware. We're now also testing for --dynamicbase and --high-entropy-va.

In addition, newer ld shows things like "--[disable-]dynamicbase", so
adjust the check to only look for the important part, then prepend "--"
to the option.

Also: --large-address-aware is only supported on 32-bit, and
--high-entropy-va is only supported on 64-bit, so restrict those.
(Technically, --large-address-aware does something on 64-bit, but
it's rather useless. If not set, it *might* restrict process memory
to 2 GB.)