[librptexture] FileFormat: Add rescale dimensions.
GodotSTEX now uses the internal dimensions as the texture size, and
sets rescale dimensions to the "custom" size if they don't match the
internal dimensions.

New image flag IMGPF_RESCALE_RFT_DIMENSIONS_2. If set, the UI frontend
should check for the second RFT_DIMENSIONS field and, if found, rescale
the image to that size first before doing anything else.

TODO: Implement this in the UI frontends.