[rpcli] Add a 'skip internal images' option for faster processing.
[librpbase] Related changes in the TextOut files.

Various textures are slow to decode, so this lets us skip the internal
image processing when simply checking the other file data.

TODO: Add something that presents just the image metadata instead of
parsing the entire image. This would allow us to keep the internal
image display without the overhead of decoding the image.