[librptexture] GodotSTEX: Handle ETCn and PVRTC images that have sizes that aren't powers of 2.
PVRTC images are always internally powers of 2, though unlike ETCn, Godot
doesn't seem to set rescaling parameters. We'll need to set that manually.
NOTE: PVRTC-II supports NPOT, but we're using a slightly-modified PVRTC-I
decoder, so it doesn't support NPOT.

calcImageSizePVRTC_PoT(): New function to calculate the PVRTC image size,
adjusting for powers of 2.

Godot 3's PVRTC textures are handled the same as ETCn with regards to
image dimensions (they're stretched out to the full size), but the rescale
values aren't set. Added a hack to handle this like ETCn so the texture is
displayed correctly.