[gtk] OptionsMenuButton: Rework into a container for GtkMenuButton.
GTK4's GtkMenuButton is a final class, so we can't subclass it.

Previously, I added a copy of the class definition from GTK in-line,
which "worked", but is subject to breakage with any GTK4 update.

The new method makes OptionsMenuButton a container (GtkBox/GtkHBox)
with signal routing for "clicked" and "activate".

RomDataView: GtkDialog connects to "activate" instead of "clicked"
because OptionsMenuButton is no longer a subclass of GtkButton.

gtk-compat.h: Added pre-GTK4 compatibility function wrappers:
- gtk_widget_class_set_activate_signal()
- gtk_widget_class_get_activate_signal()