[libromdata] MegaDrive: Verify the Mega CD security program to determine the region code.
A good number of Mega CD titles have a region code in the ROM header that
doesn't match the security program, which resulted in the wrong region
code being displayed, and the wrong region code was used to fetch the
title screen image. The security program determines the *actual* system
region, so verify that to get the real region.

We're checking the CRC of the first 0x150 bytes of the security program.
This is the same on all JPN discs, all USA discs, and all EUR discs.
Note that USA and EUR have the same first 0x4C0 bytes, but we can't rely
on that for JPN, so we'll use the smaller amount.

NOTE: imgdb has been updated to use the new arrangement, so some Mega CD
title screens won't show up on older versions of rom-properties anymore.