[libromdata] MegaDrive: Handle multi-disc title screens for "Slam City with Scottie Pippen".
The disc number is in the MCD-specific header, though this isn't a
standard thing. Other multi-disc games either have something different
or don't have the disc number at all.

If the serial number has 'F' in the last position, assume it's 32X.
The European version has "SEGA 32X" in the header, but the USA version
has "SEGA GENESIS", which prevented it from being detected properly.