[libromdata] CBMCart: Add external image downloading.
C64 carts don't have identifiable titles or serial numbers.
(The title field in VICE's .CRT format isn't reliable.)

Instead, we'll use the CRC32 of the first 16 KB of data in the
cartridge. This requires parsing the CHIP headers to identify the
actual data. This allows us to get matching CRC32s for different
.CRT files if the ROM data is the same but the headers aren't.

NOTE: *Up to* 16 KB is used. For 8 KB ROMs, 8 KB is used.