[kde] RomThumbCreator: Work around an issue in KDE's thumbnailer that breaks if img.bytesPerLine() is larger than width * bytespp.
This only seems to show up with hi_mark_sq.ktx, which is 145x130
24-bit RGB.

To work around it, make a copy of the QImage using QImage::copy().
This allocates a new memory buffer which uses "generic" alignment.

TODO: Figure out why it breaks in KIO's thumbnailer and submit a fix
upstream.