[gtk] RomDataView: Clean up tab widget deletion.
_RomDataViewCxx::tab: Get rid of the destructor and `page` field.

rom_data_view_init_string(): Actually set tab.lblCredits.

rom_data_view_delete_tabs(): If a single tab is in use, delete the
table (GtkTable/GtkGrid) before clearing the tabs vector. If more
tabs are in use, deleting the tabWidget (GtkNotebook) will handle
this for us, so we don't have to delete anything ourselves.

Use gtk_container_remove() [GTK+ 2.x/3.x] or gtk_box_remove() [GTK4]
to delete widgets instead of gtk_widget_destroy() or g_object_unref().
g_object_unref() in particular doesn't work right because it causes
the object to be deleted, but the parent object still holds a reference
to it, so things get out of sync.

TODO: OptionsMenuButton still uses gtk_widget_destroy(), but it's in a
code path that isn't built in GTK4 builds, so it shouldn't be a problem.