[gtk] OptionsMenuButton: Remove the std::unordered_map<> for actions.
I originally tried to do this on 2021/08/06 07:57 PM EDT, but ran into
an issue where the renamed menu items were always disabled. Turns out
this was because when calling g_menu_insert() to re-add the menu item,
the detailed_action parameter was set to the ID only, without the
menu button prefix. The GAction in the actionGroup only uses the ID
as its name, but when added to a GMenu, it needs the full prefix.

This should save a bit of memory, since we don't need an extra
std::unordered_map<> anymore.