[libromdata] LuaPrivate::reset_lua(): Don't reset luaVersion.
reset_lua() is called by parse(). The Windows frontend ends up
calling the parse() function multiple times, which causes version
to be reset (it's set when the file is loaded, not when it's parsed),
which results in an error in systemName because the index is out of
bounds. (Explorer then crashes.)

Also, change LuaVersion back to a standard int because int8_t was
being converted to 0xFF instead of 0xFFFFFFFF (-1).

This bug was reported by @HyperBE32 .