[librpbase] RpImageLoader: Convert from a static class to a namespace.
FIXME: For some reason, ImageDecoderTest isn't able to find the
RpImageLoader::load() function, even when RP_LIBROMDATA_PUBLIC is
added to the declaration. Converting to a namespace didn't fix this,
but it allowed for some code simplification.

NOTE: This only caused problems when we stopped linking the .a libraries
directly, which is necessary to fully deduplicate the code when using
libromdata as a shared library.