[librpfile] CMakeLists.txt: Mark librpfile as an OBJECT library.
This ensures that VectorFile doesn't get truncated when linking
libromdata as a shared library.

NOTE: CMake was failing after this change because rpfile and rptexture
had circular dependencies, and this is only allowed for STATIC libraries.
I'm not sure why rptexture was listed as a dependency (it isn't needed),
so I removed it, along with inih.