[libromdata] AmiiboData, ExecRpDownload_win32.cpp: Don't use dll_filename from libromdata.
dll_filename would need to be exported from the Win32 UI frontend, which
creates a circular reference when libromdata is converted to a shared
library. It isn't really needed anyway, since we can just call
GetModuleFileName() whenever we need it.

AmiiboData: Don't store the DLL filename in a C++ string, since
we can easily retrieve it using GetModuleFileName().