[gtk] KeyStoreGTK: Initial GTK+ implementation.
KeyStoreGTK is a GObject, so it can't directly inherit from KeyStoreUI.
KeyStoreGTK.cpp has an internal C++ class that derives from KeyStoreUI
to forward signals, and it provides a function to get the underlying
KeyStoreUI in order to get key management data.

Currently, it adds the key names to the GtkTreeView.

TODO: Section names, values, editing, "Valid?".