Add separate directories for the GTK3 and GTK4 versions of LibNautilusExtension.
Since we don't actually need the LibNautilusExtension package, I removed
the FIND_PACKAGE() calls and hard-coded the paths.

On my test system, I had the GTK4 plugin installed in the Nautilus GTK3
directory, since that's where rp-config (libunixcommon) was looking for
it. Nautilus attempted to load *both* plugins, which failed miserably.
Splitting the directories and updating libunixcommon to handle this
fixes Nautilus.