Refactored RomOps handlers to use a table for Standard Actions information.
The table contains the Save dialog title, file filter, default file
extension, and whether or not the action is copying to the clipboard.
Note that if it *is* copying to the clipboard, everything else is NULL,
which might waste a bit of space, but it's generally more efficient than
the previous setup.

Code size differences: (64-bit Gentoo Linux, gcc-12.1.0, release build, no LTO)

   text    data     bss     dec     hex filename
  31776      72      89   31937    7cc1 [GTK3] RomDataView.cpp.o [before]
  31626     168      89   31883    7c8b [GTK3] RomDataView.cpp.o [after]
   -150     +96       0     -54     -36 Difference

   text    data     bss     dec     hex filename
  34378       8       0   34386    8652 [KDE] RomDataView.cpp.o [before]
  34062     104       0   34166    8576 [KDE] RomDataView.cpp.o [after]
   -316     +96       0    -220     -dc Difference