[libromdata] ndscrypt.cpp: Use unique_ptr<> for blowfish_data.
Encrypting and decrypting Nintendo DS ROMs is a rarely-done operation,
so we shouldn't allocate 0x1048 * 3 == 4,168 * 3 == 12,504 bytes (in
the .bss section) unless it's actually needed.

The array of std::unique_ptr<> takes up 32 bytes, according to bloaty.
There's one unique_ptr for each Blowfish key, so if only regular NDS
ROM images are needed, then only 4,168 bytes will be allocated instead
of the full 12,504 bytes.

NOTE: std::unique_ptr<> should be memory-safe because the destructor
is automatically called when the shared library is unloaded.