[gtk] KeyManagerTab: Implemented key editing.
The key is set in the KeyStoreGTK, which then emits signals to tell the
UI to update the key text and the Reset/Apply buttons.

Removed the 'inhibit' variable since it isn't needed here.

TODO: Actually save the keys.

[libromdata] KeyStoreUI: setKey() now takes `const char*` instead of
`const std::string&`. Also modified convertKanjiToHex(). This does
change the LibRomData ABI, but since we haven't finalized a build
using LibRomData as a shared library yet, that isn't a problem.

[win32] KeyManagerTab: Updated for the KeyStoreUI::setKey() change.

NOTE: KeyStoreQt was apparently passing a `const char*` to setKey()
already, so that was implicitly being converted to an std::string.