[libromdata] CMakeLists.txt: Explicitly link to -lgcc when compiling with gcc.
Xubuntu 16.04 fails to link libromdata to anything due to an issue with
gcc's internal CPUID functions:

/usr/bin/ld: ../../../bin/TextFuncsTest: hidden symbol `__cpu_indicator_init'
	in /usr/lib/gcc/i686-linux-gnu/5/libgcc.a(cpuinfo.o) is referenced by DSO
/usr/bin/ld: final link failed: Bad value
collect2: error: ld returned 1 exit status
src/librpbase/tests/CMakeFiles/TextFuncsTest.dir/build.make:107:
	recipe for target 'bin/TextFuncsTest' failed
make[2]: *** [bin/TextFuncsTest] Error 1

This function is in libgcc.a, which is statically-linked to executables,
and is *supposed* to be linked to libromdata.so, but it isn't for some
reason. However, it works fine on my main system, which has gcc-12.1.

Adding -lgcc makes it work everywhere.