Explicitly link to -lgcc if using gcc-5 or earlier.
rom-properties-xfce.so was also missing libgcc:

Thunar :Failed to load plugin `rom-properties-xfce.so':
	/usr/lib/i386-linux-gnu/thunarx-2/rom-properties-xfce.so:
	undefined symbol: __cpu_model

This seems to be a bug with gcc-5 in general, so if using gcc-5 or
earlier, link everything to libgcc. Since it's a static library,
only symbols that are actually used will be included.

Note that unlike the unit test linker errors, this one only showed up
when attempting to load the plugin in Thunar.