[xfce] CMakeLists.txt: Add ${GTK2_GTK_LIBRARY_RELEASE}.
For some reason, FindGTK2.cmake on Xubuntu 16.04 (cmake-3.5.1) isn't
adding libgtk-x11-2.0.so to ${GTK2_LIBRARIES}, though I'm pretty sure
it did before. This resulted in rom-properties-xfce.so linking to all
of the necessary GTK+ libraries *except* for libgtk-x11-2.0.so, so it
couldn't even find gtk_init().

TODO: Now that the GTK2 rp-config loads, fix all the warnings and
GUI layout issues.