[libromdata] KeyStoreUI: Consolidate the import functions into a single dispatch function.
This removes four very-specific functions from the exported symbol table
and replaces it with a more generic function that can be expanded without
breaking the ABI later on.

Moved the GTK+ KeyManagerTab's ImportMenuID to KeyStoreUI and renamed it
to ImportFileID, and turned it into an enum class.

TODO: [kde,win32] KeyManagerTab: Rewrite the "Import" menu action handlers
to use a dispatch function, similar to the GTK+ version.